Extract PlayerList
